DO NOT MERGE: Make recovery permissive
This is to make it easier for people to build TWRP in the CM tree
if for some reason they desire to do so. Some devices wil not obey
the kernel command line option (exynos, omap) so I am pushing this
patch so that they can easily cherry pick it. I will abandon it
shortly.

Change-Id: I14b9104ba26c3725fa51d32355d917d5f025944a